requirements.txt
torch
dgl
numpy
networkx
POT
faiss-gpu